QuantTrade_AI/requirements.txt
# Core Numerical & Data Libraries
numpy>=2.0,<3.0
pandas>=2.2,<2.3
scipy>=1.13,<1.14

# Machine Learning & State-Space Modeling
scikit-learn>=1.5,<1.6
yfinance>=0.2.30,<0.3.0

# Backend Framework & Server
fastapi>=0.109,<0.120
uvicorn[standard]>=0.27,<0.32
pydantic>=2.8,<2.9
pydantic-settings>=2.2,<2.3
python-multipart>=0.0.6,<0.1.0

# Database & Vector Search
SQLAlchemy>=2.0,<2.1
psycopg2-binary>=2.9,<2.10
pgvector>=0.2.4,<0.3.0
sentence-transformers>=2.5.1,<3.0.0
requests>=2.31,<2.33

# Utilities
pyarrow>=16.0,<17.0
python-dotenv>=1.0,<2.0
